server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.7.0
google-cloud-storage==2.19.0
google-cloud-firestore==2.20.0
google-cloud-pubsub==2.27.1
python-dotenv==1.0.1
